add `scenario_single_client` testing utility to reduce boilerplate
