+++ crates/valence_new/src/unit_test/example.rs
@@ -1,9 +1,8 @@
 use bevy_app::App;
 
-use super::util::create_mock_client;
 use crate::config::ServerPlugin;
 use crate::server::Server;
-use crate::unit_test::util::gen_client_info;
+use crate::unit_test::util::scenario_single_client;
 
 /// Examples of valence unit tests that need to test the behavior of the server,
 /// and not just the logic of a single function. This module is meant to be a
@@ -17,7 +16,6 @@ mod tests {
 
     use super::*;
     use crate::client::Client;
-    use crate::dimension::DimensionId;
     use crate::inventory::{Inventory, InventoryKind, OpenInventory};
 
     /// The server's tick should increment every update.
@@ -38,10 +36,7 @@ mod tests {
     #[test]
     fn example_test_client_position() {
         let mut app = App::new();
-        app.add_plugin(ServerPlugin::new(()));
-        let info = gen_client_info("test");
-        let (client, mut client_helper) = create_mock_client(info);
-        let client_ent = app.world.spawn(client).id();
+        let (client_ent, mut client_helper) = scenario_single_client(&mut app);
 
         // Send a packet as the client to the server.
         let packet = valence_protocol::packets::c2s::play::SetPlayerPosition {
@@ -62,19 +57,10 @@ mod tests {
     #[test]
     fn example_test_open_inventory() -> anyhow::Result<()> {
         let mut app = App::new();
-        app.add_plugin(ServerPlugin::new(()));
-        let server = app.world.resource::<Server>();
-        let instance = server.new_instance(DimensionId::default());
-        let instance_ent = app.world.spawn(instance).id();
-        let info = gen_client_info("test");
-        let (mut client, mut client_helper) = create_mock_client(info);
-        client.set_instance(instance_ent); // HACK: needed so client does not get disconnected on first update
+        let (client_ent, mut client_helper) = scenario_single_client(&mut app);
+
         let inventory = Inventory::new(InventoryKind::Generic3x3);
         let inventory_ent = app.world.spawn(inventory).id();
-        let client_ent = app
-            .world
-            .spawn((client, Inventory::new(InventoryKind::Player)))
-            .id();
 
         // Process a tick to get past the "on join" logic.
         app.update();
